pyproject.toml
[project]
name = "mnist"
version = "0.4.0"
description = "Default template for PDM package"
authors = [
    {name = "WhiteCapella", email = "elnath1002@naver.com"},
]
dependencies = [
    "fastapi>=0.114.1",
    "uvicorn[standard]>=0.30.6",
    "PyMySQL>=1.1.1",
    "pytz>=2024.2",
    "pillow>=10.4.0",
    "python-multipart>=0.0.9",
    "tensorflow>=2.17.0",
    "keras>=3.5.0",
    "matplotlib>=3.9.2",
    "scikit-learn>=1.5.2",
    "scikit-image>=0.24.0",
    "opencv-python>=4.10.0.84",
    "pydot>=3.0.1",
]

requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
note = [
    "notebook>=7.2.2",
]
test = [
    "pytest>=8.3.3",
]

